Move url helpers into media-url-utils
