+++ src/utils/media-url-utils.js
@@ -68,6 +68,14 @@ export const proxiedUrlFor = url => {
   }
 };
 
+export function getAbsoluteUrl(baseUrl, relativeUrl) {
+  return new URL(relativeUrl, baseUrl);
+}
+
+export function getAbsoluteHref(baseUrl, relativeUrl) {
+  return getAbsoluteUrl(baseUrl, relativeUrl).href;
+}
+
 export const getCustomGLTFParserURLResolver = gltfUrl => url => {
   if (typeof url !== "string" || url === "") return "";
   if (/^(https?:)?\/\//i.test(url)) return url;
